% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{best_10_winpct}
\alias{best_10_winpct}
\title{-----Top teams by win percentage}
\usage{
best_10_winpct(data)
}
\arguments{
\item{data}{a data.frame with W et G columns}
}
\value{
ordered tibble by win percentage
}
\description{
A function returning top 10 teams by win percentage
}
