% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{filter_conference}
\alias{filter_conference}
\title{------Filter conference}
\usage{
filter_conference(data, conf = "ACC")
}
\arguments{
\item{data}{dataset cbb}

\item{conf}{the desired conference (ex: "ACC","B12","SEC",...)}
}
\value{
tibble with teams from the selected conference
}
\description{
A function filtering the dataset by conference
}
