% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_top_conference}
\alias{plot_top_conference}
\title{Plot conference ranking based on a selected metric}
\usage{
plot_top_conference(data, metric = "ADJOE")
}
\arguments{
\item{data}{A dataframe containing at least two columns:
\code{CONF} (conference names) and the selected metric.}

\item{metric}{A character string indicating which metric to plot c("ADJOE","ADJDE","EFG_O","EFG_D")}
}
\value{
A \code{ggplot} object representing a horizontal bar chart.
}
\description{
This function creates a simple horizontal bar chart showing the average
value of a selected performance metric for each NCAA conference.
It is designed to be used with the output of \code{top_conference()},
but works with any dataframe containing \code{CONF} and the chosen metric.
}
