% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{team_profil}
\alias{team_profil}
\title{------Team profile}
\usage{
team_profil(data, team)
}
\arguments{
\item{data}{dataset cbb}

\item{team}{any of the 365 ncaa teams the 2025-26 season}
}
\value{
polygon chart
}
\description{
A function retunring a polygon plot of the chosen team overall performance
}
