% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{top_defense}
\alias{top_defense}
\title{------Top defense by metrics}
\usage{
top_defense(data, n = 10, metric = c("ADJDE", "EFG_D", "TORD"))
}
\arguments{
\item{data}{dataset cbb}

\item{n}{number of teams}

\item{metric}{options of offensive metrics between ("ADJDE","EFG_D","TORD")}
}
\value{
ordered tibble
}
\description{
A function returning the best teams by defensive metric chosen
}
