% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{top_offense}
\alias{top_offense}
\title{------Top offense by metrics}
\usage{
top_offense(data, n = 10, metric = c("ADJOE", "EFG_O", "2P_O", "3P_O"))
}
\arguments{
\item{data}{dataset cbb}

\item{n}{number of teams}

\item{metric}{options of offensive metrics between ("ADJOE","EFG_O","2P_O","3P_O")}
}
\value{
ordered tibble
}
\description{
A function retunring the best teams by offensive metric chosen
}
