% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{top_conference}
\alias{top_conference}
\title{-------Top conference}
\usage{
top_conference(
  data,
  metric = c("ADJOE", "ADJDE", "EFG_O", "EFG_D"),
  desc = TRUE
)
}
\arguments{
\item{data}{dataset cbb}

\item{metric}{the desired metric for the ranking from this list:c("ADJOE","ADJDE","EFG_O","EFG_D")}

\item{desc}{boolean TRUE/FALSE, TRUE by default,}
}
\value{
tibble with the 31 conferences ranked
}
\description{
A function to rank conference by desired metrics
}
